Building with libusb git repo
Forked libusb git repo modified for android is to be used instead of
the downloaded archive.
Also, the patch is removed as it is included in git repo.
libusb is modified to use file descriptor for making a libusb_device.
Maybe this will help get past the permission issue. The plan is to
get user permission for usage of USB device and then get the USB
devices fd and make a usb device from that.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>