Initial JNI Implementation for USB Permission
This is an early attempt to get the required permissions for usage of usb
devices on android via native code. It implements verbatim all
the steps required for getting the usb permissions from Home.java.
As can be seen, the JNI code is quite lengthy.
There are many shortcomings as of now.
1. No listing of the USB device. This code chooses the first FTDI device
   it can get its hands on. And asks for its permission.
2. No visible indication of progress. You only get a blank screen.
3. This native code still relies on Java - UsbPermissionReceiver
   class. As android does not allow creating of classes from the native
   code, it was not possible to extend a BroadcastReceiver from the
   native code. There might be a workaround somehow. It is not even
   necessary. But the application crashes in the absence of
   BroadcastReceiver.
4. No exception handling as of yet. This is just a crude initial
   implementation.

After getting USB permissions, the application starts downloading
dives from OSTC 3. The output can be seen in logcat.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>