APP_ABI not recognised by Application.mk
For some reason, jni/Application.mk recognizes variables that are
defined by ndk-build. Hence APP_ABI, even though it looked like
it was working for arm, but in actuality, it wasn't recognised. The
default option is arm. Hence, it worked for arm. But compiling for x86
architecture, it still used arm-linux-androideabi-gcc.
Hence using TARGET_ARCH_ABI.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>